novel_to_video_mvp/requirements.txt
openai
moviepy
edge-tts
requests
python-dotenv
asyncio
pydub
